Experimenting further with Pixi
Copied function's code from his branch, src/components/match-pixi/main.tsx
Mainly did this so that I can understand how the map is generated from the tile images in the repo.
Basically, each tile is a sprite added on top of a pixi Application and that application is connected to the canvas return by the component.

Next:
Analyze the weird black background when the page is loaded
Look into loading the tile from a sprite sheet